capstone_project/requirements.txt
fastapi
uvicorn
sentence-transformers
faiss-cpu
groq
python-dotenv
pydantic
PyPDF2
pytest
pytest-cov
httpx
